StructuredType simple implementations for List and Map
